+++ backend-manager-student/src/admin_api/v1/controllers/admin.controller.js
@@ -1,6 +1,3 @@
-const admin_service = require('../../../share/services/admin.service');
-const helper = require('../../../share/utils/helper');
-
 const adminController = {
     /**
      * @author Nguyễn Tiến Tài
@@ -13,27 +10,11 @@ const adminController = {
     LoginAdmin: async (req, res) => {
         const { user_name, password } = req.body.input.admin_login_input;
         try {
-            const ip = req.headers['x-forwarded-for'] || req.socket.remoteAddress || null;
-
-            const { device } = req;
-
-            const geo = helper.findingLocationByIP(ip) || helper.findingLocationByIP('14.165.41.226');
-            console.info(geo, '-------------', device);
-            const data = await admin_service.createAdmin('oke');
-            if (data !== 'oke') {
-                return res.status(400).json({
-                    status: 400,
-                    message: 'error',
-                });
-            }
             return res.status(200).json({
                 status: 200,
                 data: {
                     user_name,
                     password,
-                    geo,
-                    ip,
-                    device,
                 },
                 message: 'success',
             });
